requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
exceptiongroup==1.2.0
executing==2.0.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
openai==1.3.4
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==23.3.1
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.2
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.2
tiktoken==0.5.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
wheel==0.41.3
zipp==3.17.0
